add:messages for user controller

+++ src/main/java/fr/codechill/spring/rest/UserController.java
@@ -26,6 +26,9 @@ import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.node.ObjectNode;
+
 import fr.codechill.spring.controller.DockerController;
 import fr.codechill.spring.exception.BadRequestException;
 import fr.codechill.spring.model.Docker;
@@ -62,56 +65,98 @@ public class UserController {
     }
 
     @GetMapping("/user/{id}")
-    public User getUser(@PathVariable("id") Long id) {
-        User user = this.urepo.findOne(id);
-        logger.info(user);
-        logger.info("getting user informations");
-        return user;
+    public ResponseEntity<?> getUser(@PathVariable("id") Long id) {
+
+        ObjectMapper mapper = new ObjectMapper();
+        HttpHeaders headers = new HttpHeaders();
+        ObjectNode body = mapper.createObjectNode();
+        if (this.urepo.findOne(id)!=null)
+        {
+            User user = this.urepo.findOne(id);
+            body.set("user",this.constructJsonUser(user));
+            return ResponseEntity.ok().headers(new HttpHeaders()).body(body);
+        }
+        body.put("message","no user found");
+        return ResponseEntity.badRequest().headers(headers).body(body);
     }
 
 
     @DeleteMapping("/user")
     public ResponseEntity<?> deleteUser(@RequestHeader(value="Authorization") String token) {
         String username = jwtTokenUtil.getUsernameFromToken(token.substring(7));
-        User user = this.urepo.findByUsername(username);
-        this.urepo.delete(user);
-        logger.info("the user"+ user.getLastname()+"has been deleted");
-        return ResponseEntity.ok().headers(new HttpHeaders()).body(null);
+        ObjectMapper mapper = new ObjectMapper();
+        HttpHeaders headers = new HttpHeaders();
+        ObjectNode body = mapper.createObjectNode();
+        if (this.urepo.findByUsername(username)!=null)
+        { 
+            User user = this.urepo.findByUsername(username);
+            this.urepo.delete(user);
+            body.put("message "," the user "+ user.getLastname()+" has been deleted");
+            logger.info(body.toString());
+            return ResponseEntity.ok().headers(new HttpHeaders()).body(body);
+        }
+        else
+        {
+            body.put("message","the user your trying to delete doesn't seems to exist :shrug");
+            return ResponseEntity.badRequest().headers(headers).body(body);
+        }
     }
 
     @PutMapping("/user")
-    public User editUser(@RequestHeader(value="Authorization") String token, @RequestBody User user) {
-
+    public ResponseEntity<?> editUser(@RequestHeader(value="Authorization") String token, @RequestBody User user) {
+        ObjectMapper mapper = new ObjectMapper();
+        HttpHeaders headers = new HttpHeaders();
+        ObjectNode body = mapper.createObjectNode();
         String username = jwtTokenUtil.getUsernameFromToken(token.substring(7));
         User updatedUser = this.urepo.findByUsername(username);
+
         if (!updatedUser.getLastname().equals(user.getLastname())) {
             logger.info("updating user lastname with : "+updatedUser.getLastname());
             updatedUser.setLastname(user.getLastname());
+            body.put("message","user's lastname updated");
+            body.set("user",this.constructJsonUser(user));
+            logger.info("body "+ body.toString());
+            this.urepo.save(updatedUser);
+            return ResponseEntity.ok().headers(headers).body(body);
         }
+
         if (!updatedUser.getFirstname().equals(user.getFirstname())) {
             logger.info("updating user lastname with : "+updatedUser.getFirstname());
             updatedUser.setFirstname(user.getFirstname());
+            body.put("message","user's firstname updated");
+            body.set("user",this.constructJsonUser(user));
+            this.urepo.save(updatedUser);
+            return ResponseEntity.ok().headers(headers).body(body);
             
         }
+
         if (!updatedUser.getEmail().equals(user.getEmail())) {
             updatedUser.setEmail(user.getEmail());
             logger.info("updating user email with : "+ updatedUser.getEmail());
-            updateUserEmail(updatedUser.getEmail());         
+            updateUserEmail(updatedUser.getEmail());
+            body.put("message","user's email updated");
+            body.set("user",this.constructJsonUser(user));
+            this.urepo.save(updatedUser);
+            return ResponseEntity.ok().headers(headers).body(body);         
         }
-        this.urepo.save(updatedUser);
-        return user;
+        body.put("message","the user your trying to edit doesn't seem to exist :shrug");
+        return ResponseEntity.badRequest().headers(headers).body(body);
     }
 
     @PostMapping("/user")
     public ResponseEntity<?> addUser(@RequestBody User user) throws BadRequestException {
+        ObjectMapper mapper = new ObjectMapper();
+        ObjectNode body = mapper.createObjectNode();
         HttpHeaders responseHeaders = new HttpHeaders();
         if (urepo.findByUsername(user.getUsername()) != null) {
             logger.info("An account with this username already exist ");
-            throw new BadRequestException("An account with this username already exist!");
+            body.put("message","An account with this username already exist,returning a bad request");
+            return ResponseEntity.badRequest().headers(responseHeaders).body(body);
         }
         if (urepo.findByEmail(user.getEmail()) != null) {
-            logger.info("An account with this email already exist!");
-            throw new BadRequestException("An account with this email already exist!");
+            logger.info("An account with this email already exist,returning a bad request");
+            body.put("message","An account with this email already exist!");
+            return ResponseEntity.badRequest().headers(responseHeaders).body(body);
         }
         user.setPassword(bCryptPasswordEncoder.encode(user.getPassword()));
         Docker docker = this.dcontroller.createDocker();
@@ -120,7 +165,8 @@ public class UserController {
         final URI location = ServletUriComponentsBuilder
             .fromCurrentServletMapping().path("/user/{id}").build()
             .expand(user.getId()).toUri();
-        return ResponseEntity.created(location).headers(responseHeaders).body(urepo.findByUsername(user.getUsername()));
+        body.set("user",this.constructJsonUser(user));
+        return ResponseEntity.created(location).headers(responseHeaders).body(body);
     }
 
 
@@ -162,8 +208,9 @@ public class UserController {
 	@PostMapping(value = "/user/forgottenpassword")
 	public ResponseEntity<?> processForgotPasswordForm(@RequestBody String email) throws BadRequestException {
         User user = urepo.findByEmail(email);
+        ObjectMapper mapper = new ObjectMapper();
+        ObjectNode body = mapper.createObjectNode();
         HttpHeaders responseHeaders = new HttpHeaders();
-
         if(user != null) {
             user.setLastPasswordResetDate(new Date());
             user.setTokenPassword(UUID.randomUUID().toString());
@@ -174,16 +221,20 @@ public class UserController {
             passwordResetEmail.setTo(user.getEmail());
             passwordResetEmail.setSubject("Password Reset Request");
             passwordResetEmail.setText("Reset link:\n" + BASE_URL + "/reset/" + user.getTokenPassword());
-
             mailSender.send(passwordResetEmail);
-            return ResponseEntity.ok().headers(responseHeaders).body(user);
+            body.put("message","an email has been sent to reset the password");
+            body.set("user",this.constructJsonUser(user));
+            return ResponseEntity.ok().headers(responseHeaders).body(body);
         }
         logger.info("No user email was found with this email");
-        throw new BadRequestException("No user found with this email");  
+        body.put("message","No user email was found with this input");
+        return ResponseEntity.badRequest().headers(responseHeaders).body(body);  
     }
 
     @GetMapping(value = "/reset/{token}")
     public ResponseEntity<?> resetPassword(@PathVariable("token") String token) throws BadRequestException {
+        ObjectMapper mapper = new ObjectMapper();
+        ObjectNode body = mapper.createObjectNode();
         HttpHeaders responseHeaders = new HttpHeaders();
         User user = urepo.findByTokenPassword(token);
         Date currentDate = new Date();
@@ -192,24 +243,45 @@ public class UserController {
         c.add(Calendar.DATE, 1);
         Date currentDatePlusOne = c.getTime();
         if(user != null && currentDate.after(user.getLastPasswordResetDate()) && currentDate.before(currentDatePlusOne)) {
-            return ResponseEntity.ok().headers(responseHeaders).body(user);
+            body.put("message","the password has been reset");
+            body.set("user",this.constructJsonUser(user));
+            return ResponseEntity.ok().headers(responseHeaders).body(body);
         }
         logger.info("Password reset failed due to an invalid token or an out dated one");
-        throw new BadRequestException("Your token is invalid or hax expired");
+        body.put("message","Your token is invalid or hax expired");
+        return ResponseEntity.badRequest().headers(responseHeaders).body(body);  
     }
 
     @PostMapping(value = "/reset")
     public ResponseEntity<?> setNewPassword(@RequestBody Map<String, String> requestParams) throws BadRequestException {
         User user = urepo.findByTokenPassword(requestParams.get("token"));
+        ObjectMapper mapper = new ObjectMapper();
+        ObjectNode body = mapper.createObjectNode();
         HttpHeaders responseHeaders = new HttpHeaders();
 
         if(user != null) {
             user.setPassword(bCryptPasswordEncoder.encode(requestParams.get("password")));
             user.setTokenPassword(null);
             urepo.save(user);
-            return ResponseEntity.ok().headers(responseHeaders).body(user);
+            body.put("message","a new password has been set");
+            body.set("user",this.constructJsonUser(user));
+            return ResponseEntity.ok().headers(responseHeaders).body(body);
         }
         logger.info("Setting of a new password due to an invalid token or an out dated one");
-        throw new BadRequestException("Your token is invalid or hax expired");
+        body.put("message","Your token is invalid or hax expired");
+        return ResponseEntity.badRequest().headers(responseHeaders).body(body);  
+    }
+
+    private ObjectNode constructJsonUser(User user) {
+        ObjectMapper mapper = new ObjectMapper();
+        ObjectNode jsonUser = mapper.createObjectNode();
+        jsonUser.put("ID",user.getId());
+        jsonUser.put("Username",user.getUsername());
+        jsonUser.put("Password",user.getPassword());
+        jsonUser.put("Firstname",user.getFirstname());
+        jsonUser.put("Lastname",user.getLastname());
+        jsonUser.put("email",user.getEmail());
+        logger.info("constructed user : " + jsonUser.toString());
+        return jsonUser;
     }
 }
